Add previewImage to schema

+++ app/services/session.server.ts
@@ -81,6 +81,9 @@ export async function getUser(request: Request) {
 
 export async function logout(request: Request) {
   const session = await getUserSession(request);
+  if (!session) {
+    return redirect("/");
+  }
   return redirect("/login", {
     headers: {
       "Set-Cookie": await sessionStorage.destroySession(session),
